UI Completed with dark-light mode
