+++ lib/main.dart
@@ -1,9 +1,7 @@
-// dart package
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:voice_search/screens/home_screen.dart';
 import 'package:voice_search/utils/app_strings.dart';
-
-// local package
-import './custom_video_controllers.dart';
 
 void main() {
   runApp(const MyApp());
@@ -14,14 +12,33 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: AppStrings.voiceSearch,
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
+    return ChangeNotifierProvider<ThemeModel>(
+      create: (context) => ThemeModel(), // Initialize the theme model
+      child: Consumer<ThemeModel>(
+        builder: (context, theme, child) {
+          return MaterialApp(
+            debugShowCheckedModeBanner: false,
+            title: AppStrings.voiceSearch,
+            theme: theme.isDark
+                ? ThemeData.dark(useMaterial3: true)
+                : ThemeData(
+                    useMaterial3: true,
+                  ), // Use the selected theme
+            home: const HomeScreen(),
+          );
+        },
       ),
-      home: const CustomVideoControllers(),
     );
+  }
+}
+
+class ThemeModel with ChangeNotifier {
+  bool _isDark = false; // Default to light theme
+
+  bool get isDark => _isDark;
+
+  void toggleTheme() {
+    _isDark = !_isDark;
+    notifyListeners(); // Notify listeners to update the UI
   }
 }
