+++ pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   avatar_glow: ^2.0.2
   speech_to_text: ^6.3.0
+  provider: ^6.0.5
 
 
 dev_dependencies:
